requirements.txt
google-generativeai>=0.3.0
pillow>=9.0.0
pandas>=1.5.0
python-dotenv>=0.19.0
tqdm>=4.64.0
pathlib>=1.0.0 